:building_construction: cleanup code and re-structure main using functions from mods
